Catch file exceptions and throw DBErrors instead.
If the database file could not be accessed (in method connect()) or be
created (in connect() and create()) the methods raised
FileNotFoundErrors. Those are caught now and DBErrors with a appropriate
message are raised. This is to prevent the user from submitting bugs,
that occurred, because the module is used wrong.