Store the ID in every row.
A new method select() was added. This method is similar to read(), but
the returned list consists out of tuples instead of the rows. The tuples
are formed out of (id, row).
The other methods, like read(), work as before.

Note that the ID is currently the length of the table, which works for
now, but if a method delete() is introduced this has to change, or
multiple IDs will occur.