Implemented delete().
Added tests.
Added a member variable that holds the highest given ID for every table.
Load the highest ID form a file and restore the tuples, that are gone
missing during the save to and restore from JSON.
Setup highest ID in _create_db().
Use highest ID in insert() instreal of the length of the table.
Added delete(). This method checks, if the ID is in the table at all,
and raises an exception if not. After that the whole table is iterated
through and every occurrence of the ID (which should be exactly one) is
removed.